First prototype
For now, you can switch between long and wide versions, but number of slots is
fixed and there's only one drawer. The goal is to allow multiple drawers (grow
downwards) and allow longer versions. something like this:

          +-------+
          |       |
     +---+|       |+---+
     | o ||       || x |
     |   ||       ||x x|
     | x ||       || x |
     |x x||       ||   |
     | x ||       || o |
     +---+|       |+---+
          |       |
          +-------+

Of course it gets obnoxious large combinations, but what's the point of doing
something parametric if not to allow insane scenarios? :)

Signed-off-by: Aristeu Rozanski <aris@ruivo.org>